refactor: remove models to project
