+++ .docker/entrypoint.sh
@@ -2,4 +2,6 @@
 
 yarn install
 
+cp .env.example .env
+
 yarn start:dev
