--- src/domain/models/company.ts
@@ -1,8 +1,0 @@
-export interface CompanyModel {
-  id: string
-  name: string
-  cnpj?: string
-  logo?: string
-  celPhone?: string
-  createdAt?: Date
-}
