--- src/domain/models/employee.ts
@@ -1,11 +1,0 @@
-import { type CompanyModel } from './company'
-
-export interface EmployeeModel {
-  id: string
-  name: string
-  email: string
-  password: string
-  company: string | CompanyModel
-  avatar?: string
-  createdAt?: Date
-}
