--- src/presentation/controllers/login/signup/signup-protocols.ts
@@ -1,5 +1,0 @@
-export * from '../../../protocols'
-export * from '../../../protocols/email-validator'
-export * from '../../../../domain/usecases/employee/add-employee'
-export * from '../../../../domain/usecases/employee/authentication'
-export * from '../../../../domain/usecases/company/add-company'
