+++ src/presentation/protocols/controller.ts
@@ -1,10 +1,5 @@
-import { type HttpRequest, type HttpResponse } from './http'
+import { type HttpResponse } from './http'
 
-export interface Controller<Q, R> {
-  handle: (HttpRequest: Controller.Request<Q>) => Promise<Controller.Response<R>>
-}
-
-export namespace Controller {
-  export type Request<Q> = HttpRequest<Q>
-  export type Response<R> = HttpResponse<R>
+export interface Controller<T = any> {
+  handle: (request: T) => Promise<HttpResponse>
 }
