+++ src/presentation/protocols/http.ts
@@ -1,7 +1,3 @@
-export interface HttpRequest<B = any> {
-  body?: B
-}
-
 export interface HttpResponse<B = any> {
   statusCode: number
   body: B
