+++ tests/domain/mocks/mock-company.ts
@@ -1,9 +1,4 @@
-import { type AddCompany } from '../../../src/domain/usecases/company/add-company'
-import { type CompanyModel } from '../../../src/domain/models/company'
-export const mockCompany = (): CompanyModel => ({
-  id: 'any_id',
-  name: 'any_name'
-})
+import { type AddCompany } from '../usecases'
 
 export const mockAddCompanyParams = (): AddCompany.Params => ({
   email: 'any_mail@mail.com',
