+++ tests/presentation/mocks/mock-company.ts
@@ -1,10 +1,11 @@
-import { type AddCompany } from '../../../src/domain/usecases/company/add-company'
-import { mockCompany } from '../../domain/mocks'
+import { type AddCompany } from '../../domain/usecases'
 
 export const AddCompanySpy = (): AddCompany => {
   class AddCompanySpy implements AddCompany {
     async add (company: AddCompany.Params): Promise<AddCompany.Result> {
-      return await Promise.resolve(mockCompany())
+      return await Promise.resolve({
+        id: 'any_id'
+      })
     }
   }
