+++ tests/presentation/mocks/mock-signup.ts
@@ -1,5 +1,5 @@
-import { type SignUpController } from '../../../src/presentation/controllers/login/signup/signup'
-export const mockSignUpControllerRequest = (): SignUpController.Request => ({
+import { type SignUpController } from '../controllers/signup'
+export const mockSignUpControllerRequest = (): SignUpController.Params => ({
   email: 'any_mail@mail.com',
   name: 'any_name',
   password: 'any_password',
